requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
burnysc2==7.1.1
frozenlist==1.7.0
idna==3.10
loguru==0.7.3
mpyq==0.2.5
multidict==6.6.3
numpy==2.3.2
portpicker==1.6.0
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
s2clientprotocol==5.0.14.93333.0
scipy==1.16.1
typing-extensions==4.14.1
yarl==1.20.1
